% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\docType{data}
\name{AcousticWeightingTable}
\alias{AcousticWeightingTable}
\title{Weighting acoustic table}
\description{
Weighting acoustic table
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
\keyword{data}
