% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{AcuPercentile}
\alias{AcuPercentile}
\title{Calculate reverse Percentile}
\usage{
AcuPercentile(x)
}
\arguments{
\item{x}{is a vector with Leq data}
}
\description{
Returns a vector of acoustic percetile
}
\examples{
# Calculate reverse quantile of a vector of values

#data(PTFA)

AcuPercentile(PTFA$LAeq)
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
