% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\docType{data}
\name{dataset_impulsive1}
\alias{dataset_impulsive1}
\title{Noise dataset of impulsive event (100 ms acquisition time)}
\description{
Noise dataset of impulsive event (100 ms acquisition time)
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
\keyword{data}
