% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{dbsum}
\alias{dbsum}
\title{Calculate energetic sum or difference of values}
\usage{
dbsum(x, y, operator)
}
\arguments{
\item{x}{is first value or vector}

\item{y}{is second value or vector}

\item{operator}{is 1 for sum and -1 for difference}
}
\description{
Returns energetic sum or difference of values
}
\examples{
# Calculate energetic sum or difference of values

dbsum(x = 55, y = 33, operator = 1)
dbsum(x = c(55 , 66), y = c(45, 50), operator = 1)

dbsum(x = c(70 , 68), y = c(55, 66), operator = -1)
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
