% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{deco.time}
\alias{deco.time}
\title{Time decomposition}
\usage{
deco.time(x, y, z, verbose = TRUE)
}
\arguments{
\item{x}{are hours}

\item{y}{are minutes}

\item{z}{are seconds}

\item{verbose}{logic argument that on or off message (default is TRUE)}
}
\value{
time decomposition in seconds
}
\description{
Trasform time from hours, minutes and seconds to seconds
}
\examples{
# Convert H:M:S to seconds
x = 5  # Hours
y = 25 # minutes
z = 50 # seconds

deco.time(x = x, y = y, z = z)
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
