% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\docType{data}
\name{dfBW}
\alias{dfBW}
\title{Table's 1/3 octave bandwidth}
\description{
Table's 1/3 octave bandwidth
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
\keyword{data}
