% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{dfImpulsiveTrasform}
\alias{dfImpulsiveTrasform}
\title{Trasform impulsive dataframe (100 ms samples) in dataframe (1s samples)}
\usage{
dfImpulsiveTrasform(dfImpulsive, statistic = energetic.mean)
}
\arguments{
\item{dfImpulsive}{is a dataframe for impulse}

\item{statistic}{is energetic mean (default)}
}
\description{
Returns a dataframe (1s samples)
}
\examples{
#data("dataset_impulsive1")
#data("dataset_impulsive2")

dfImpulsiveTrasform(dfImpulsive = dataset_impulsive1,
                    statistic = energetic.mean)[1:5, ]
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
