% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{energetic.mean}
\alias{energetic.mean}
\title{Logarithmic  mean}
\usage{
energetic.mean(x)
}
\arguments{
\item{x}{is a vector of value in decibel (dB)}
}
\value{
logarithmic mean
}
\description{
Calculate logarithmic mean
}
\examples{
# Calculate energetic mean of vector's values of Leq
energetic.mean(c(55, 88, 66, 51, 70))
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
