% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{energetic.min}
\alias{energetic.min}
\title{Function that calculate min value}
\usage{
energetic.min(y)
}
\arguments{
\item{y}{is a numeric vector}
}
\description{
calculate min value
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
