% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{energetic_w.mean}
\alias{energetic_w.mean}
\title{Weigth logaritmic  mean}
\usage{
energetic_w.mean(x, t)
}
\arguments{
\item{x}{is a vector of value in decibel (dB)}

\item{t}{is a vector of time string "HH:MM:SS"}
}
\value{
weigth logarithmic mean  respect to time
}
\description{
Calculate weigth logarithmic mean respect to time
}
\examples{
# Calculate weight energetic mean
energetic_w.mean(x = c(55.0, 70.0) , t = c("03:55:22", "01:33:12"))
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
