% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\docType{data}
\name{exampleHourlyData}
\alias{exampleHourlyData}
\title{Noise hourly data of misure in environmental open space}
\description{
Noise hourly data of misure in environmental open space
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
\keyword{data}
