% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{ExtractIndexMark}
\alias{ExtractIndexMark}
\title{Extract index and name of markers}
\usage{
ExtractIndexMark(filemarks, dataset, mp)
}
\arguments{
\item{filemarks}{is a dataframe with date and markers}

\item{dataset}{is dataframe in analysis}

\item{mp}{is a name of misure point}
}
\value{
list of index and names
}
\description{
Returns a list of index and name
}
\examples{
#data(PTFA)
#data("markers")

ExtractIndexMark(filemarks = markers , dataset = PTFA, mp = "PTFA")
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
