% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{HolidaysDate}
\alias{HolidaysDate}
\title{Calculate Holidays date (Gregorian calendar)}
\usage{
HolidaysDate(year_holiday)
}
\arguments{
\item{year_holiday}{is year example "2022" like character}
}
\description{
Returns a vector of holiday dates (Gregorian calendar)
}
\examples{
HolidaysDate(2022)
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
