% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{HourlyEmean}
\alias{HourlyEmean}
\title{Calculate hourly energetic mean}
\usage{
HourlyEmean(df, variable, timeZone = "Europe/Rome")
}
\arguments{
\item{df}{is a dataframe with date (Y-m-d H:M:S) and variables}

\item{variable}{is a variable name}

\item{timeZone}{is time zone defoult is Europe/Rome}
}
\description{
Returns a dataframe with hourly energetic mean
}
\examples{
#data(PTFA)

HourlyEmean(df = PTFA, variable = "LAeq")
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
