% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\docType{data}
\name{iso}
\alias{iso}
\title{Parameters table of equal loudness curve A (ISO 226:1987 “Acoustics -- Normal equal-loudness-level contours”)}
\description{
Parameters table of equal loudness curve A (ISO 226:1987 “Acoustics -- Normal equal-loudness-level contours”)
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
\keyword{data}
