% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\docType{data}
\name{markers}
\alias{markers}
\title{Dataset with markers}
\description{
Dataset with markers
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
\keyword{data}
