% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{Maskapply}
\alias{Maskapply}
\title{Add index and name of markers in misure dataframe}
\usage{
Maskapply(filemarks, dataset, mp)
}
\arguments{
\item{filemarks}{is a dataframe with date and markers}

\item{dataset}{is dataframe in analysis}

\item{mp}{is a name of misure point}
}
\description{
Returns a dataframe
}
\examples{
#data(PTFA)
#data(markers)

Maskapply(filemarks = markers, dataset = PTFA, mp = "PTFA")[1:10, c(1, 2, 45)]
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
