% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\docType{data}
\name{P1FC}
\alias{P1FC}
\title{Noise data of misure in house close window condition}
\description{
Noise data of misure in house close window condition
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
\keyword{data}
