% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{PlotNoiseTHcompare}
\alias{PlotNoiseTHcompare}
\title{Plot time histrory and compare frequency components}
\usage{
PlotNoiseTHcompare(
  df,
  variable,
  listvar = NULL,
  mp,
  runleq = TRUE,
  y_lim = c(20, 80)
)
}
\arguments{
\item{df}{is a dataframe}

\item{variable}{is Leq or another variable to plot first}

\item{listvar}{are names of frequency component you want compare}

\item{mp}{is a misure point}

\item{runleq}{is logical value that plot running leq line}

\item{y_lim}{y axe range}
}
\description{
Returns a plot
}
\examples{
#data(P1FA)

PlotNoiseTHcompare(df = P1FA ,
                   variable = "LAeq",
                   listvar = c("LZFmin.100",
                               "LZFmin.250"), mp = "P1FA",
                   runleq = TRUE,
                   y_lim = c(30, 70))
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
