% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{PlotSpectrogram}
\alias{PlotSpectrogram}
\title{Plot spectrogram}
\usage{
PlotSpectrogram(df, coLs, plot_title = NULL)
}
\arguments{
\item{df}{is a dataframe}

\item{coLs}{is cols index to plot}

\item{plot_title}{is title of plot}
}
\description{
Returns a spectrogram
}
\examples{
#data(P1FA)

PlotSpectrogram(df = P1FA, coLs = c(3:38) , plot_title = "Spettrogram LZFmin")
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
