% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{RoundTo}
\alias{RoundTo}
\title{Round to Multiple}
\usage{
RoundTo(x, multiple = 1, FUN = round)
}
\arguments{
\item{x}{is a vector of value in decibel (dB)}

\item{multiple}{is a vector of time in seconds}

\item{FUN}{the rounding function as character or as expression. Can be one out of trunc, ceiling, round (default) or floor.}
}
\description{
Returns a number rounded to the nearest specified multiple.
}
\examples{
#data("P1FA")

RoundTo(x = P1FA$LAeq, multiple = 0.5)[1:10]
}
\author{
Andri Signorell \email{andri@signorell.net}
}
