% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{runningLeq}
\alias{runningLeq}
\title{Calculate running Leq}
\usage{
runningLeq(x)
}
\arguments{
\item{x}{is a vector of value in decibel (dB)}
}
\description{
Returns a vector of energetic mean of Leq......
}
\examples{
#data("P1FA")

runningLeq(x = P1FA$LAeq)[1:10]
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
