% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{search.tone}
\alias{search.tone}
\title{Function research pure tone}
\usage{
search.tone(x, statistic = energetic.mean, plot.tone = FALSE)
}
\arguments{
\item{x}{is a dataframe with llfmin...}

\item{statistic}{is statistic used default is energetic.mean}

\item{plot.tone}{is logic argument default is false don't plot result}
}
\value{
plot of 1/3 octave frequency and isofonic curve A (ISO 226:1987)
}
\description{
research pure tone
}
\examples{
# Search pure tone from PTFA dataset
# dataframe have to contain date and LLfmin (minor value of Linear level frequency)
search.tone(PTFA[, c(3:38)], plot.tone = FALSE)

# Plot result
search.tone(PTFA[, c(3:38)], plot.tone = TRUE)
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
