% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{searchImpulse}
\alias{searchImpulse}
\title{Search impulsive event}
\usage{
searchImpulse(df, cri1 = 6, cri2 = -10, Threshold = 30)
}
\arguments{
\item{df}{is a impulse dataframe, samples of 100 ms}

\item{cri1}{is first criteria 6dB (LAImax - LASmax > 6dB)}

\item{cri2}{is second criteria -10dB ((LAFmax - 10dB) < 1s)}

\item{Threshold}{is minimun level for detect peaks}
}
\value{
list contain a dataframe of peaks values and a plot of it
}
\description{
Returns a list with dataframe of peaks impulsive and a plot
}
\examples{
#data("dataset_impulsive1")
#data("dataset_impulsive2")

searchImpulse(df = dataset_impulsive1)
searchImpulse(df = dataset_impulsive2)
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
