% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{SELcalc}
\alias{SELcalc}
\title{Calculate SEL (Single Event Level)}
\usage{
SELcalc(x, t)
}
\arguments{
\item{x}{is value in dB}

\item{t}{is period in second}
}
\description{
Returns SEL
}
\examples{
# Calculate SEL (Single Event Level)

SELcalc(x = 66.8, t = 938)
}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
