% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpeNoise.R
\name{IntrusiveIndex}
\alias{IntrusiveIndex}
\title{Calculate Intrusive Index (UNI/TS 11844 march 2022)}
\usage{
IntrusiveIndex(dfa, dfr, BW)
}
\arguments{
\item{dfa}{is a dataframe Lfa (enviromental sound levels) 1/3 octave specta data}

\item{dfr}{is a dataframe Lfr (residual sound levels) 1/3 octave specta data}

\item{BW}{a vector of 1/3 octave bandwidth data}
}
\description{
Returns a number
}
\examples{

}
\author{
Pasquale Scordino \email{p.scordino@arpa.piemonte.it}

Simone Sperotto \email{s.sperotto@arpa.piemonte.it}
}
